Fix: converge dependencies (#34)
This PR should solve some dependency convergence issues for groovy libraries. Also, the data model lib version for the project has been pinned in the root pom, the modules only reference it. Like this, we can ensure that the data model version is same in the modules all the time.