Remove unnecessary parameters from binding methods
This commit removes the unnecessary parameters from binding the view model
and listening to the view model. Here the instance fields are used and
no parameters are needed.

This commit changes the UI design slightly. The toggle button always displays
the current status and an icon was added to the create customer button.