Feature/create affiliation fields (#56)
* Remove unnecessary parameters from binding methods

This commit removes the unnecessary parameters from binding the view model
and listening to the view model. Here the instance fields are used and
no parameters are needed.

This commit changes the UI design slightly. The toggle button always displays
the current status and an icon was added to the create customer button.

* Provide empty fields with validation for createAffiliation

* Remove unused code

* Bind view model for createAffiliation and provide validation

This commits checks on user input change if the fields are marked as valid.
When all fields are valid the button is enabled. Otherwise it is disabled.

* Add overlooked arguments to CreateAffiliationController method

* Add items to combo boxes

* Update Changelog

* Add missing binding to affiliations list