Copy method from CustomerDatabaseQueries
In order to prepare testablility and simplify the code the DatabaseQueries and the CustomerDbConnector can be merged.
The separation into two classes was not object driven and did not add to the code except for the need of an additional interface.

The idea is to provide the constructor of CustomerDbConnector with a java.sql.Connection obtained through DatabaseSession.groovy.