Fix/#612 refactor dependency manager by context (#624)
This PR restructures the DependencyManager. The components are now created in separate methods based on their context. A separate creation method allows for the creation of each component accessible from the AppView.

* Use ResourcesService interface instead of concrete classes

* Use DataSource interfaces instead of concrete classes outside the DB connection method

* Introduce missing interfaces between service and database connectors 

This does not introduce interfaces between database connectors but only between the service part and the database part

* Extract component construction into helper methods

* Update Changelog

* Change OfferOverviewView creation

For the OfferOverviewView to work the way the CreateProjectView was constructed
was changed. This change was necessary because the CreateProjectView is only
relevant for the OfferOverviewView. There must be a new CreateProjectView for every
OfferOverviewView to ensure that the component is only added there.

* Improve JavaDoc

Co-authored-by: jnnfr <jennifer.boedker@qbic.uni-tuebingen.de>

* Add missing since tag

Co-authored-by: jnnfr <jennifer.boedker@qbic.uni-tuebingen.de>